readded version.
By accident I broke your deployment process.

Due to the missing concrete version in the pom maven will attempt to resolve the parent to inherit the version. This will happen on every build/dependency update. Since you are not deploying the parent to central (for good reason) this breaks on the client machines (it probably only works on your machine because of your local repository).

I readded the version and please do a quick redeployment of this thing :) Thanks and sorry
Jonas